requirements.txt
# Core Framework
streamlit>=1.28.0
streamlit-code-editor>=0.1.0

# Elasticsearch Integration
elasticsearch>=8.10.0

# LLM & AI Libraries
anthropic>=0.7.0
openai>=1.3.0
langchain>=0.1.0
langchain-anthropic>=0.0.1

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
faker>=19.0.0  # For synthetic data generation
plotly>=5.18.0  # For interactive visualizations

# Web Search & Research
httpx>=0.25.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utilities
tenacity>=8.2.3  # For retry logic
tqdm>=4.66.0
python-json-logger>=2.0.7

# File Generation
jinja2>=3.1.2  # For template generation
markdown>=3.4.4

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
ruff>=0.0.280

# Type hints
types-requests>=2.31.0
mypy>=1.4.0eland[pytorch]>=8.0.0

